Add files via upload
This is a face recognition software i started working on January 23, 2019. I took this up as a project for my final year defense and as a way to learn about machine learning and python programming language. It can be implemented across schools, and universities where attendance is required. I took up this project as my final year project to be defended at the end of my 4 year bachelor degree in Redeemer's University. I am constantly improving on the system and will be pushing more files here.